refactor: create new hook for atc actions
